pyproject.toml
[project]
name = "it-guru"
version = "0.1.0"
description = "Тестовое задание для компании IT Guru - https://docs.google.com/document/d/1B5ibBRpdu08bouBbtPg5UcCySeht0xyXNOMqYocQvLo/edit?tab=t.0"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]",
    "tortoise-orm[asyncpg]",
    "aerich",
    "pydantic-settings",
    "httpx",
    "pytest>=9.0.2",
    "celery[redis]",
    "flower",
]

[dependency-groups]
dev = [
    "faker",
    "pytest",
    "pytest-asyncio",
    "ruff",
    "pre-commit",
    "httpx",
]

[tool.aerich]
tortoise_orm = "src.database.TORTOISE_ORM"
location = "./src/migrations"
src_folder = "."

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py313"
line-length = 120

[tool.ruff.lint]
select = [
    "E",
    "F",
    "W",
    "I",
    "N",
    "UP",
    "B",
    "SIM",
]
ignore = ["B008"]

[tool.ruff.lint.per-file-ignores]
"src/migrations/**" = ["E501"]
